requirements.txt
altgraph==0.17
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
future==0.18.2
get==2019.4.13
idna==2.8
numpy==1.18.1
opencv-python==4.1.2.30
pandas==0.25.3
pefile==2019.4.18
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
PyInstaller==3.6
pytesseract==0.3.1
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
query-string==2019.4.13
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.7
